MIDAS MK2.1 Port -- final software release v3

+++ MIDAS/src/silsim/emulated_telemetry.cpp
@@ -9,7 +9,7 @@ ErrorCode __attribute__((warn_unused_result)) TelemetryBackend::init() {
     return ErrorCode::NoError;
 }
 
-int8_t TelemetryBackend::getRecentRssi() {
+int16_t TelemetryBackend::getRecentRssi() {
     return 1;
 }
